+++ NEWS.md
@@ -1,5 +1,7 @@
 # CHANGES IN xfun VERSION 0.49
 
+- Added an argument `use_block = FALSE` to `protect_math()`. When `use_block = TRUE`, a `$$ $$` expression that spans across multiple lines will be protected in a code block.
+
 # CHANGES IN xfun VERSION 0.48
 
 - Added utilities for HTML tags: `html_tag()`, `html_escape()`, `html_escape()`, and `html_view()`. Removed the soft dependency on the **htmltools** package accordingly.